fix: keep wait absent coverage gates green
