+++ src/mcp/__tests__/command-tools-input-docs.test.ts
@@ -1,0 +1,167 @@
+import assert from 'node:assert/strict';
+import { test } from 'vitest';
+import { listCommandTools } from '../command-tools.ts';
+
+// Guidance no longer restates input fields in prose, so a tool's inputSchema is the only
+// place its inputs are documented — for the model, for `--help`, and for the docs site. An
+// undescribed property is therefore a silent gap rather than a cosmetic one.
+//
+// The baseline pins exact `tool.property` identities, not bare property names plus a total:
+// a name-and-count baseline stays green when a gap migrates (describe `foo.text`, add an
+// undescribed `bar.text`, and both the allowed-name set and the total are unchanged), and
+// stale names keep authorizing gaps that appear later. Exact identities make every new or
+// moved gap fail, and require deleting an entry to record a fix.
+const UNDESCRIBED_TOOL_INPUTS = new Set([
+  'alert.action',
+  'alert.timeoutMs',
+  'audio.action',
+  'audio.probeAction',
+  'back.mode',
+  'clipboard.action',
+  'clipboard.text',
+  'close.saveScript',
+  'debug.action',
+  'diff.depth',
+  'diff.interactiveOnly',
+  'diff.kind',
+  'diff.out',
+  'diff.raw',
+  'diff.scope',
+  'events.cursor',
+  'events.limit',
+  'find.action',
+  'find.depth',
+  'find.first',
+  'find.last',
+  'find.locator',
+  'find.query',
+  'find.raw',
+  'find.timeoutMs',
+  'find.value',
+  'get.format',
+  'install-from-source.retainPaths',
+  'install-from-source.retentionMs',
+  'is.predicate',
+  'is.selector',
+  'is.value',
+  'keyboard.action',
+  'logs.action',
+  'logs.message',
+  'logs.restart',
+  'metro.action',
+  'metro.bridgeScope',
+  'metro.bundleUrl',
+  'metro.installDependenciesIfNeeded',
+  'metro.kind',
+  'metro.launchUrl',
+  'metro.listenHost',
+  'metro.logPath',
+  'metro.metroHost',
+  'metro.metroPort',
+  'metro.port',
+  'metro.probeTimeoutMs',
+  'metro.projectRoot',
+  'metro.publicBaseUrl',
+  'metro.reuseExisting',
+  'metro.runtimeFilePath',
+  'metro.startupTimeoutMs',
+  'metro.statusHost',
+  'metro.timeoutMs',
+  'network.action',
+  'network.include',
+  'network.limit',
+  'open.saveScript',
+  'orientation.orientation',
+  'perf.action',
+  'perf.area',
+  'perf.kind',
+  'perf.subject',
+  'push.app',
+  'push.payload',
+  'react-native.action',
+  'record.action',
+  'record.fps',
+  'record.hideTouches',
+  'record.path',
+  'record.quality',
+  'record.recordingScope',
+  'reinstall.app',
+  'replay.backend',
+  'replay.env',
+  'replay.force',
+  'replay.maestro',
+  'replay.path',
+  'replay.resumeFrom',
+  'replay.resumePlanDigest',
+  'replay.saveScript',
+  'replay.update',
+  'screenshot.fullscreen',
+  'screenshot.normalizeStatusBar',
+  'screenshot.overlayRefs',
+  'screenshot.stabilize',
+  'screenshot.surface',
+  'scroll.direction',
+  'settings.app',
+  'settings.latitude',
+  'settings.longitude',
+  'settings.mode',
+  'settings.permission',
+  'settings.setting',
+  'settings.state',
+  'snapshot.depth',
+  'snapshot.forceFull',
+  'snapshot.interactiveOnly',
+  'snapshot.raw',
+  'snapshot.scope',
+  'swipe.pattern',
+  'test.artifactsDir',
+  'test.backend',
+  'test.env',
+  'test.failFast',
+  'test.maestro',
+  'test.paths',
+  'test.recordVideo',
+  'test.retries',
+  'test.shardAll',
+  'test.shardSplit',
+  'test.timeoutMs',
+  'test.update',
+  'trace.action',
+  'trace.path',
+  'tv-remote.button',
+  'wait.depth',
+  'wait.durationMs',
+  'wait.absent',
+  'wait.kind',
+  'wait.quietMs',
+  'wait.raw',
+  'wait.ref',
+  'wait.scope',
+  'wait.selector',
+  'wait.stable',
+  'wait.text',
+  'wait.timeoutMs',
+]);
+
+test('MCP tool inputs do not add undocumented properties', () => {
+  const undescribed: string[] = [];
+  for (const tool of listCommandTools()) {
+    for (const [key, schema] of Object.entries(tool.inputSchema.properties ?? {})) {
+      if (!schema.description) undescribed.push(`${tool.name}.${key}`);
+    }
+  }
+
+  const added = undescribed.filter((entry) => !UNDESCRIBED_TOOL_INPUTS.has(entry)).sort();
+  assert.deepEqual(
+    added,
+    [],
+    `These MCP tool inputs need a schema description: ${added.join(', ')}`,
+  );
+
+  const fixed = [...UNDESCRIBED_TOOL_INPUTS].filter((entry) => !undescribed.includes(entry)).sort();
+  assert.deepEqual(
+    fixed,
+    [],
+    `These MCP tool inputs are documented now — remove them from UNDESCRIBED_TOOL_INPUTS: ${fixed.join(', ')}`,
+  );
+});
